% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bizcompany.R
\name{bizcompany}
\alias{bizcompany}
\title{Create a dataframe of company and branch level data}
\usage{
bizcompany(n)
}
\arguments{
\item{n}{int, sample size of dataframe}
}
\value{
dataframe
}
\description{
Generate a dataframe of size integer n of company level data
}
\examples{
bizcompany(10)
bizcompany(150)
}
