% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bizstaff.R
\name{bizstaff}
\alias{bizstaff}
\title{Create a dataframe of random employee level data}
\usage{
bizstaff(n)
}
\arguments{
\item{n}{int, sample size of dataframe}
}
\value{
dataframe
}
\description{
Generate a dataframe of size integer n of employee level data
}
\examples{
bizstaff(10)
bizstaff(150)
}
