% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testmean.R
\name{testmean}
\alias{testmean}
\title{Mean of a number}
\usage{
testmean(a)
}
\arguments{
\item{a}{vector}
}
\value{
float
}
\description{
Takes the mean by using mean function
}
\examples{
testmean(c(3,4,5))
}
